move plugin related functionality into `IdeaPlugin.kt`
