+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/LinChecker.kt
@@ -110,7 +110,6 @@ class LinChecker(private val testClass: Class<*>, options: Options<*, *>?) {
                     minimizedScenario.run(this, createVerifier())
                 }
             }
-            reporter.logFailedIteration(failure)
             runReplayForPlugin(failure, verifier)
             return failure
         }
@@ -124,7 +123,6 @@ class LinChecker(private val testClass: Class<*>, options: Options<*, *>?) {
      */
     private fun CTestConfiguration.runReplayForPlugin(failure: LincheckFailure, verifier: Verifier) {
         if (ideaPluginEnabled() && this is ModelCheckingCTestConfiguration) {
-            reporter.logFailedIteration(failure, loggingLevel = LoggingLevel.WARN)
             enableReplayModeForIdeaPlugin()
             val strategy = createStrategy(failure.scenario)
             check(strategy is ModelCheckingStrategy)
@@ -133,8 +131,6 @@ class LinChecker(private val testClass: Class<*>, options: Options<*, *>?) {
                 check(replayedFailure != null)
                 strategy.runReplayIfPluginEnabled(replayedFailure)
             }
-        } else {
-            reporter.logFailedIteration(failure)
         }
     }
